float type fix

+++ python_introduction/finance_calculator.py
@@ -1,7 +1,7 @@
 user_income = int(input("Enter your monthly income:"))
 total_monthly_expenses = int(input("Enter your total monthly expenses:"))
-monthly_savings = user_income - total_monthly_expenses
-projected_savings = monthly_savings * 12 + (monthly_savings * 12 * 0.05)
+monthly_savings = float(user_income - total_monthly_expenses)
+projected_savings = float(monthly_savings * 12 + (monthly_savings * 12 * 0.05))
 
 print(f"Your monthly savings are ${monthly_savings}.")
 print(f"Projected savings after one year, with interest, is:  ${projected_savings}.")
